% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcaAnalysis.r
\name{ascoreSeperation}
\alias{ascoreSeperation}
\title{The Processing of PC values
A utility designed to help seperate a vector based on the mean and standard deviation of the data. For the most part ascoreSeparation will return a vector of logicals but funs such as "value" and "rank" return numerics.}
\source{
\url{https://github.com/alexjgriffith/mulcal}
}
\usage{
ascoreSeperation(vect, fun, n = 1)
}
\arguments{
\item{vect}{A vector of numerical values.}

\item{fun}{A function which is used to sepearate memebers of the vector.
\describe{
\item{\strong{"top"}}{TRUE if greater than mean + n* standard deviation.}
\item{\strong{"bottom"}}{TRUE if less than mean - n* standard deviation.}
\item{\strong{"nottop"}}{TRUE if less than mean + n* standard deviation.}
\item{\strong{"notbottom"}}{TRUE if greater than mean - n* standard deviation.}
\item{\strong{"middle"}}{TRUE if within mean +/- n* standard deviation.}
\item{\strong{"value"}}{Returns the principle component of interest.}
\item{\strong{"rank"}}{Returns the ranks of the principle components.}
}}

\item{n}{The value which is used to adjust the standard deviation cut off.}
}
\value{
A vector of equal length to the parameter vect.
}
\description{
The Processing of PC values
A utility designed to help seperate a vector based on the mean and standard deviation of the data. For the most part ascoreSeparation will return a vector of logicals but funs such as "value" and "rank" return numerics.
}
\author{
Alexander Griffith griffita@gmail.com
}

