% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/motifAnalysis.r
\name{compliment}
\alias{compliment}
\title{reverse compliment}
\source{
\url{https://github.com/alexjgriffith/mulcal}
}
\usage{
compliment(string)
}
\description{
Takes a string which can include base nucleaotides (AGCT) and brakets ("[","]") and returns the reverse compliment including brakets
}
\author{
Alexander Griffith griffita@gmail.com
}

