% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heights.r
\name{convertChroms}
\alias{convertChroms}
\title{convertChroms}
\source{
\url{https://github.com/alexjgriffith/mulcal}
}
\usage{
convertChroms(lis)
}
\arguments{
\item{lis}{a list of chromosmes to}
}
\description{
Calls valueChromosome from hg19Height.c which converts a chromosme rank to a string
}
\author{
Alexander Griffith griffita@gmail.com
}
\seealso{
For the reverse see \code{\link{rankChroms}}
}

