% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rnaSeqAnalysis.r
\name{exonsfromGenome}
\alias{exonsfromGenome}
\title{Exons from Genome}
\usage{
exonsfromGenome(genome, chroms = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13,
  14, 15, 16, 17, 18, 19, 20, 21, 22))
}
\description{
Generates a bed file format for exons from a RefSeqGene set.
}
\examples{
genome<-read.table("hg19.RefSeqGenes.csv",header=TRUE)
exonList<-exonsFromGenome(genomes)
write.table(exonLocs,"hg19.exonLocs",row.names=FALSE,col.names=FALSE,quote=FALSE,sep="\\t")
}

