% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gene-assoc.r
\name{geneAssociation}
\alias{geneAssociation}
\title{Gene Association}
\source{
\url{https://github.com/alexjgriffith/mulcal}
}
\usage{
geneAssociation(bedData, geneList, bounds, n = FALSE)
}
\arguments{
\item{bedData}{three column data.frame}

\item{geneList}{a data.frame loaded without a header from the refseq gene data base}

\item{bounds}{c(A,B,C,D) Region 1 = (tss-A,tss+B) Region 2 = (ess-C ess+D)}

\item{n}{Number of nodes to use.}
}
\description{
Assosiate each ChIP enrinchement region with the closest peak within a bound defined by the user
}
\examples{
fileLocation="~/Dropbox/UTX-Alex/jan/"
bedData<-read.delim(paste(fileLocation,"combined_sorted.bed",sep=""),header=0)
geneList<-read.delim(paste(fileLocation,"hg19.RefSeqGenes.csv",sep=""))
genes<-geneAssociation(bedData,geneList,  c(50000,0,0,0))
write.table(cbind(bedData,unlist(t(lapply(genes, function(x) {if(identical(x,character(0))){"None"} else{x}})))) ,"combined_tagged_genes.bed" ,col.names=FALSE,row.names=FALSE,quote=FALSE,sep="\\t")
}
\author{
Alexander Griffith griffita@gmail.com
}

