% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gene-assoc.r
\name{genomicRegions}
\alias{genomicRegions}
\title{Genomic Regions}
\usage{
genomicRegions(chrom, tss, proxUp, proxDown, distal, n = length(chrom))
}
\description{
Implements the Stanford GREAT genomic region generation
}

