% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/heights.r
\name{getData}
\alias{getData}
\title{getData}
\source{
\url{https://github.com/alexjgriffith/mulcal}
}
\usage{
getData(file, rawdata, n = 0)
}
\arguments{
\item{file}{peak file}

\item{rawdata}{a list of raw data files}

\item{n}{the number of nodes used default=0}
}
\description{
loads data from file returns a sorted set with a height matrix
}
\author{
Alexander Griffith griffita@gmail.com
}

