% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gene-assoc.r
\name{getMin}
\alias{getMin}
\title{Get Min
will return the single value if there is only a single value in the list}
\usage{
getMin(x)
}
\description{
Get Min
will return the single value if there is only a single value in the list
}

