% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heights.r
\name{getPileUp}
\alias{getPileUp}
\title{getPileUp}
\source{
\url{https://github.com/alexjgriffith/mulcal}
}
\usage{
getPileUp(file, bed, chroms, peakLength)
}
\arguments{
\item{file}{The raw data file of interest}

\item{bed}{The preloaded bed infromation including bed$start and bed$end}

\item{chroms}{a list of chromosomes whos string values have been repalced with ranks}

\item{peakLength}{The length of the bed data provided}
}
\description{
A minimal wrapper for the pileup function from hg19Height.c
returns a integer vector of computed read pileups
}
\author{
Alexander Griffith griffita@gmail.com
}

