% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/motifAnalysis.r
\name{heightHist}
\alias{heightHist}
\title{Height Histrogram
 A plotting tool for histograms, Requires GGPLOT2}
\usage{
heightHist(t0, xlab = "Histogram")
}
\description{
Height Histrogram
 A plotting tool for histograms, Requires GGPLOT2
}

