% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heights.r
\name{hg19Sort}
\alias{hg19Sort}
\title{hg19Sort}
\source{
\url{https://github.com/alexjgriffith/mulcal}
}
\usage{
hg19Sort(data)
}
\description{
Reorders the chro factor in data to that which is outputed by bwa
}
\author{
Alexander Griffith griffita@gmail.com
}

