% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/motifAnalysis.r
\name{histVisualize}
\alias{histVisualize}
\title{Visualize batches of height data}
\usage{
histVisualize(h, m1, m2, n = 1, name = c("NoName"))
}
\description{
Visualize batches of height data
}

