% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/motifAnalysis.r
\name{homerWrapper}
\alias{homerWrapper}
\title{Homer Wrapper}
\usage{
homerWrapper(sequences, foreground, background, homerLocation,
  motifsFile = FALSE, opts = "-S 25 -len 6")
}
\description{
This is only a wrapper for the DENOVO functionality of homer2, homer must be installed on the system before this can be used.
}

