% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/html.r
\name{htmlDoc}
\alias{htmlDoc}
\title{html Doc
A minamilistic compelation of functions for the generation of html}
\usage{
htmlDoc(..., doctype = "<!DOCTYPE html>", tag = "html")
}
\description{
html Doc
A minamilistic compelation of functions for the generation of html
}
\examples{
title<-htmlTags("title", "Test Page")
head<-htmlTags("head",title)
body<-htmlTags("body",htmlTags("p","This is the test page."))
htmlDoc(head,body)
}

