% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/htmlGenerate.r
\name{htmlGenerateMain}
\alias{htmlGenerateMain}
\title{html GenearateMain
Generatates an html file to visualize motifs juxtaposed from one
another}
\usage{
htmlGenerateMain(fileLocation, sets, shortNames, imageDirectory, n = 3,
  pScoreFile)
}
\description{
html GenearateMain
Generatates an html file to visualize motifs juxtaposed from one
another
}
\examples{
sets<-c("normal_not_normal.pwm", "abnormal_not_abnormal.pwm")
shortNames<-c("n_not_a","a_not_n")
fileLocation<-"~/Dropbox/temp-data/"
pScoreFile<-"/home/agriffith/Dropbox/temp-data/p-score_table_1.tab"
imageDirectory="/home/agriffith/Dropbox/temp-data/"
printFasta(fileLocation,sets,shortNames,imageDirectory)
write(htmlGenerateMain(fileLocation,sets,shortNames,imageDirectory,n=2),
      "a-score_other.html")
}

