% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/motifAnalysis.r
\name{IUPACtoBase}
\alias{IUPACtoBase}
\title{Turn IUPAC lables into basic neucleotides}
\source{
\url{https://github.com/alexjgriffith/mulcal}
}
\usage{
IUPACtoBase(char, rl = FALSE)
}
\arguments{
\item{char}{A string of IUPAC characters
c("A","G","C","T","R","Y","S","W","K","M","B","D","H","V","N")}

\item{rl}{choice to return a value
\describe{
\item{\strong{TRUE}}{Each posible string is genearated}
\item{\strong{FALSE}}{Direct translation to brakets R=[AG]}}}
}
\description{
Turn IUPAC lables into basic neucleotides
}
\author{
Alexander Griffith griffita@gmail.com
}

