% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/heights.r
\name{loadBedFile}
\alias{loadBedFile}
\title{loadBedFile}
\source{
\url{https://github.com/alexjgriffith/mulcal}
}
\usage{
loadBedFile(file)
}
\arguments{
\item{file}{the location of the bed file of interest}
}
\value{
data.frame $chro $start $end
}
\description{
Calls file_length and read_bed from hg19Height.c and returns a 3 column bed file
}
\examples{
data<-loadBedFile(file="filelocation")
}
\author{
Alexander Griffith griffita@gmail.com
}

