% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/pcaAnalysis.r
\name{loadData}
\alias{loadData}
\title{load height file}
\usage{
loadData(file = "~/masters/normal-abnormal/single_heights.bed")
}
\value{
c(stats=matrix(),data=matrix())
}
\description{
loads a tab seperated file with pasted heights of the format
<chr><start><end><height1>...<heightn>
}

