% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/htmlGenerate.r
\name{loadPWM}
\alias{loadPWM}
\title{Load PWM
This utiltiy is designed to load several PWM formats and create a uniform layout within R for analysis}
\usage{
loadPWM(fileLocation, version = "homer")
}
\description{
Load PWM
This utiltiy is designed to load several PWM formats and create a uniform layout within R for analysis
}

