% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/motifAnalysis.r
\name{locHist}
\alias{locHist}
\title{Location Histogram
A plotting tool for histograms, Requires GGPLOT2}
\usage{
locHist(t0, xlab = "Histogram", limits = c(-32, 32))
}
\description{
Location Histogram
A plotting tool for histograms, Requires GGPLOT2
}

