% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/heights.r
\name{mainProf}
\alias{mainProf}
\title{mainProf}
\source{
\url{https://github.com/alexjgriffith/mulcal}
}
\usage{
mainProf(file = "/home/agriffith/Dropbox/UTX-Alex/jan/combined_sorted.bed",
  rawdata = rep(file, 22), n = 0)
}
\description{
A simple test function for speed
}
\author{
Alexander Griffith griffita@gmail.com
}

