% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utils.r
\name{memberWrapper}
\alias{memberWrapper}
\title{Member Wrapper
Ment for bed data where the member is found in the fourth column}
\usage{
memberWrapper(bedData, lis, n = 4, sep = "-")
}
\description{
Member Wrapper
Ment for bed data where the member is found in the fourth column
}

