% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/motifAnalysis.r
\name{motifHist}
\alias{motifHist}
\title{Motif histogram}
\usage{
motifHist(data, mList, cList, locationsM, locationsC, n1, n2, reg,
  width = c(-30, 30))
}
\description{
Motif histogram
}
\examples{
values<-loadHeightFile(heightFile)
data<-values$data
reg<-ascore(data,1,"top",3)
test<-readDNAStringSet(fastaFile,use.names = TRUE)
motifs<-as.matrix(read.table("data/normal_not_abnormal_motifs"))
mList<-unlist(lapply(c(motifs,addmotifs),IUPACtoBase))
cList<-unlist(lapply(lapply(c(motifs,"CGNNGC"),IUPACtoBase),compliment))
locationsM<-lapply(mList,grep,test)
locationsC<-lapply(cList,grep,test)
l<-length(cList)
motifHist(mList,cList,locationsM,locationsC,4,l,reg)
}

