% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/motifAnalysis.r
\name{motifs2View}
\alias{motifs2View}
\title{Motifs Two View}
\usage{
motifs2View(m1, m2, reg, Sequences, nearHeights = FALSE)
}
\description{
Visualize the similarity between two motifs,
Very ineficent,
}
\examples{
FastaFile<-"~/Dropbox/UTX-Alex/jan/combined.fasta"
Sequences <- readDNAStringSet(FastaFile, "fasta")
data<-loadHeightFile("~/Dropbox/UTX-Alex/jan/combined_heights.bed")$data
reg<-as.matrix(ascore(data,c(1,1,1),c("top","bottom","middle"),c(1,1,1)))
m<-motifs2View("CANNTG","TGACCT",reg[,3],Sequences)
k<-motifs2View("CANNTG","GATAAG",reg[,1],Sequences)
t0<-unlist(lapply(seq(min(k),max(k)),function(i)length(which(k==i))))
}

