% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/motifMatrix.r
\name{motifsFromPCA}
\alias{motifsFromPCA}
\title{motifsFromPCA}
\usage{
motifsFromPCA(Sequences, data, pc, location,
  directory = "~/Masters/mulcal/inst/data/", title = paste(directory,
  paste(apply(cbind(pc, location), 1, function(x) paste(x, collapse = "_")),
  collapse = "_"), sep = ""))
}
\description{
motifsFromPCA
}
\examples{
l<-mapply(function(a,b,c){
   motifsFromPCA(Sequences,data,a,b,title=c)
},
list(1,1,3,c(3,5),c(3,7),c(3,7),7),
list("top","bottom","top",c("top","top"),c("top","bottom"),c("top","bottom"),"top"),
paste("~/Masters/mulcal/inst/data/",list("normal","abnormal","ecfc","other","hspc","meka","diff"),".pwm",sep=""))

l<-motifsFromPCA(Sequences ,data,c(1,2,3),c("top","top","bottom"))
}

