% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/motifAnalysis.r
\name{motifTest}
\alias{motifTest}
\title{Motif Testing}
\usage{
motifTest(fastaFile = "~/Dropbox/UTX-Alex/jan/combined.fasta",
  heightFile = "~/Dropbox/UTX-Alex/jan/combined_heights.bed",
  addmotifs = c("CGNNGC"))
}
\description{
Motif Testing
}

