% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{mulcal}
\alias{mulcal}
\alias{mulcal-package}
\title{Compare multiple ChIP data sets across experimental conditions}
\description{
Compare multiple ChIP data sets across experimental conditions
}
\author{
"Alexander Griffith <griffitaj@gmail.com>"
}

