% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gene-assoc.r
\name{nearestGene}
\alias{nearestGene}
\title{Nearest Gene}
\usage{
nearestGene(b, sequence, geneStart, name, peak)
}
\description{
Nearest Gene
}
\examples{
heightFile<-"~/Dropbox/UTX-Alex/jan/combined_heights.bed"
data<-loadHeightFile(heightFile)$data
geneList<-read.delim(paste(fileLocation,"hg19.RefSeqGenes.csv",sep=""))
reg<-mapply(function(pc,loc)buildRegions(data,pc,loc)[,1],
            list(1,1,3,c(3,5),c(3,7),c(3,7),7),
            list("top","bottom","top",c("top","top"),c("top","bottom"),c("top","bottom"),"top"))
a<-geneAssocFlipAux(geneList,bedData,c(50000,500,0,0))
names<-mapply(function(a,b) paste(a,b,sep="-"), as.character(geneList$name),as.character(geneList$name2))
filenames<-lapply(cbind("erythroid","t-all","ecfc","other","hspc","meka","diff"),paste, "-genes.txt",sep="")
for (i in seq(7)){
   print(filenames[[i]])
   filename<-filenames[[i]]
   sequence<-reg[,i]
   b<-lapply(a,function(x,y) list(x[[1]],x[[2]][which(x[[2]]\%in\%y)]) ,which(sequence))
   c<-nearestGene(b,sequence,as.numeric(geneList$txStart),names,bedData)
   write.table(unique(unlist(c)),filename ,col.names = FALSE,row.names=FALSE,quote=FALSE)
}
sequence<-ascore(data,1,"top",6)
filename<-"erythroid-6-genes.txt"
}

