% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/motifAnalysis.r
\name{numberIntersect}
\alias{numberIntersect}
\title{Intersect Count}
\usage{
numberIntersect(motifA, motifB, Sequences, reg = rep(TRUE, length(Sequences)))
}
\description{
Determine the number of cases which 2 motifs intersect
}

