% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tags.r
\name{outputFile}
\alias{outputFile}
\title{outputFile}
\source{
\url{https://github.com/alexjgriffith/mulcal}
}
\usage{
outputFile(peaks, chrdata, chrcats, fileOrder, width = 150, scores = FALSE)
}
\arguments{
\item{peaks}{Two column data type generated by \code{\link{tags()}}}

\item{chrdata}{Sorted input bed files all as elements as character}

\item{width}{The width which will be outputed}
}
\description{
generates the final bed format layout <chr><start><end><tag1-tag2...-tagn>
for internal use with \code{\link{tags()}}
}
\author{
Alexander Griffith griffita@gmail.com
}

