% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/motifMatrix.r
\name{partialSigMatrix}
\alias{partialSigMatrix}
\title{PartialSigMatrix}
\usage{
partialSigMatrix(Sequences, mots, reg, ranges,
  fields = unlist(Map(as.character, 1:dim(reg)[2])), addmotifs = c(),
  n = rep((length(mots) + length(addmotifs)), 2), m = dim(reg)[2],
  filename = "test")
}
\description{
The portions of saving, motif hist and sigMatrix required to
save the data in chunks
motif_1-motif_2-enviroment
option to save every x motif_2
}

