% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gene-assoc.r
\name{peakGeneRegions}
\alias{peakGeneRegions}
\title{Peak Gene Regions}
\usage{
peakGeneRegions(bedData, genes, geneList)
}
\description{
compare peak locations to gene regions generated by genomicRegions
Return a list of genes which corospond to the overlaped gene regions
}
\examples{
heightFile<-"~/Dropbox/UTX-Alex/jan/combined_heights.bed"
geneList<-read.delim(paste(fileLocation,"hg19.RefSeqGenes.csv",sep=""))
data<-loadHeightFile(heightFile)$data
reg<-mapply(function(pc,loc)buildRegions(data,pc,loc)[,1],
            list(1,1,3,c(3,5),c(3,7),c(3,7),7),
            list("top","bottom","top",c("top","top"),c("top","bottom"),c("top","bottom"),"top"))
chrom<-as.character(geneList$chrom)
tss<-as.numeric(geneList$txStart)
start<-Sys.time()
a<-genomicRegions(chrom,tss,5000,1000,50000)
Sys.time()-start
#
jurkGenes<-peakGeneRegions(bedData[reg[,2],],a,geneList)
filenames<-lapply(cbind("erythroid","t-all","ecfc","other","hspc","meka","diff"),paste, "-genes.txt",sep="")
for (i in seq(7)){
   genes<-peakGeneRegions(bedData[reg[,i],],a,geneList)
   write.table(genes,filenames[[i]],quote=FALSE,col.names=FALSE,row.names=FALSE)}
}

