% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/motifAnalysis.r
\name{PWMtoCons}
\alias{PWMtoCons}
\title{PWM to Consensus Motif}
\usage{
PWMtoCons(x)
}
\description{
A caller for ConsesnusIUPAC, Recives PWM and returns a consusensus motif
The PWMs should be loaded through the loadPWM utility.
}
\seealso{
\link{\code{loadPWM()}}
}

