% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heights.r
\name{rankChroms}
\alias{rankChroms}
\title{rankChroms}
\source{
\url{https://github.com/alexjgriffith/mulcal}
}
\usage{
rankChroms(lis)
}
\description{
Calls rankChromosome from hg19Height.c which converts a chromosme string to its equivelent rank
}
\author{
Alexander Griffith griffita@gmail.com
}
\seealso{
For the reverse see \code{\link{convertChroms}}
}

