% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/motifAnalysis.r
\name{scoreList}
\alias{scoreList}
\title{Score List}
\usage{
scoreList(k, mot, addmotifs)
}
\description{
The preliminary investigation score mechanism
scoreList(lapply(h,function(x) lapply(x,function(y)y[[2]])),mots,addmotifs)
}

