% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/motifAnalysis.r
\name{splitBlist}
\alias{splitBlist}
\title{Split by List}
\usage{
splitBlist(mstring)
}
\description{
Applied to DNA nucleotides (ACGT). Returns a list with [] removed
}
\examples{
> consensus<-"AGCT[AGCT]G"
> splitBlist(consensus)
     buf buf buf buf buf    buf
[1,] "A" "G" "C" "T" "AGCT" "G"
}

