% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/contributions.r
\name{stackedContrib}
\alias{stackedContrib}
\title{Stacked Contribution}
\source{
\url{https://github.com/alexjgriffith/mulcal}
}
\usage{
stackedContrib(data, tags, catagories, swCat = catagories, colors = FALSE,
  inord = seq(length(swCat)), steps = 40, n = 10, f = c(function(x) {   
   mean(x) + sqrt(var(x)) * n }, function(x) {     mean(x) - sqrt(var(x)) * n
  }), sum = FALSE)
}
\description{
Stacked Contribution
}
\examples{
tags<-read.table("~/Dropbox/UTX-Alex/jan/combined_sorted.bed")[,4]
catagories<-unique(unlist(lapply(levels(tags),strsplit,"-")))
values<-loadHeightFile("~/Dropbox/UTX-Alex/jan/combined_heights.bed")
data<-values$data
info<-values$info
pcs<-as.matrix(mulcal::ascore(data,c(1:10)))


swCat<-c("normal","normal","abnormal","abnormal","abnormal","abnormal","abnormal","abnormal","meka","ecfc","stem","stem")
swCat<-c("Erythroid","K562","Jurkat","CEM","RPMI","Prima5(Brand)","Prima2","Prima5","Megakaryocite","ECFC","CD133","CD34")
names(swCat)<-c("eryt","k562","jurk","cem","rpmi","tall_p1","tall_p2","tall_p3","meka","ecfc","cd133","cd34")
tSwCat<-as.character(unlist(lapply(catagories,function(x) swCat[x])))
stackedContrib(data,pcs[,1],catagories,tSwCat)
}
\author{
Alexander Griffith griffita@gmail.com
}

