% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tags.r
\name{tags}
\alias{tags}
\title{tags}
\source{
\url{https://github.com/alexjgriffith/mulcal}
}
\usage{
tags(fileList, cats, overlapWidth, outputWidth, macsCutoff = 0)
}
\arguments{
\item{cats}{c(cat1,cat2,cat3) ; each cat relates to a single file in order}

\item{overlapWidth}{The width used to genearte the overal}

\item{outputWidth}{+/- value from the estimated peak summit (twice that of \code{\link{overlapWidth}})}

\item{filelist}{c(file1,file2,file3)}
}
\description{
Recives a list of files and catagories and generates a unified set of peaks
}
\author{
Alexander Griffith griffita@gmail.com
}

