% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/test.r
\name{testFun}
\alias{testFun}
\title{Test function}
\usage{
testFun(inputList)
}
\arguments{
\item{inputList}{}
}
\description{
This function is to test the roxygen2 markup language
}
\examples{
testFun(c(1,2,3,4))
}
\author{
Alexander Griffith <griffitaj@gmail.com>
}

