% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/tags.r
\name{testTags}
\alias{testTags}
\title{testTags}
\source{
\url{https://github.com/alexjgriffith/mulcal}
}
\usage{
testTags(file = "test.bed")
}
\description{
A qick test
}
\author{
Alexander Griffith griffita@gmail.com
}

