% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tags.r
\name{testZero}
\alias{testZero}
\title{testZero}
\source{
\url{https://github.com/alexjgriffith/mulcal}
}
\usage{
testZero(value, width)
}
\arguments{
\item{value}{integer or double which cannot be less than 0}

\item{width}{integer or double}
}
\value{
a range value+width value-width where value-width >0
}
\description{
Checks if the result of subtracting width from value is zero
}
\author{
Alexander Griffith griffita@gmail.com
}

