% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/pcaAnalysis.r
\name{ascore}
\alias{ascore}
\title{Alpha Score}
\source{
\url{https://github.com/alexjgriffith/mulcal}
}
\usage{
ascore(heightFile, pc, funs, ns, normMethod = "colQn")
}
\arguments{
\item{heightFile}{Can be either a file name, a data.frame, or a matrix}

\item{pc}{The principle component of interest}

\item{funs}{funs=c("top","bottom","nottop","notbottom","middle")}

\item{ns}{An integer to adjust the SD in the fun}

\item{normMethod}{Default="colQn", use "non" if no normalization is desired}
}
\description{
Alpha Score
}
\examples{
# load a bed file of format
   # chr start end h1 ... hn
   heightFile<-"some_bed_file.bed"

   # loadHeightFile splits the first three columns into
   # stats and the remainder into data
   values<-loadHeightFile(heightFile)

   data<-values$data
   stats<-values$stats
   pc<-1
   function<-"top" # x>mean+sd*n
   n<-3
   reg<-ascore(data,pc,function,n)
   stats[reg]
}
\author{
Alexander Griffith griffita@gmail.com
}

