% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/pcaAnalysis.r
\name{batchscore}
\alias{batchscore}
\title{Batch alpha score seperation}
\usage{
batchscore(pcs, scores, funs, ns)
}
\description{
Apply a serise of options to ascoreSeperation and return a matrix. Used by \code{\link{ascore()}}.
}
\seealso{
\code{\link{ascoreSeperation}}

\code{\link{ascore}}
}

