% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gene-assoc.r
\name{genomeDomains}
\alias{genomeDomains}
\title{Genomic Regions}
\usage{
genomeDomains(chrom, tss, strand, proxUp, proxDown, distal, n = length(chrom))
}
\description{
Implements the Stanford GREAT genomic region generation
}

