% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/heights.r
\name{pileUp}
\alias{pileUp}
\title{PileUp}
\source{
\url{https://github.com/alexjgriffith/mulcal}
}
\usage{
pileUp(data, rawdata, n = 0)
}
\arguments{
\item{data}{A preloaded bed data.frame which includes slots $chro $start $end}

\item{rawdata}{a list of raw data files}

\item{n}{the number of nodes to use. If 0 then the parrallel package is not used}
}
\description{
Generates a pile up matrix from a unified set of peaks and a list of raw data sets
}
\examples{
cats<-read.table("/home/griffita/Dropbox/UTX-Alex/jan/catagories")
   prefix<-"/mnt/brand01-00/mbrand_analysis/data_sets/"
   suffix<-"_sorted.bed"
   rawdata<-apply(cats,1,function(x){paste(prefix,x,"/",x,suffix,sep="")})
   data<-hg19Sort(loadBedFile(file))
   score<-pileUp(data,rawdata,n=22)
   temp<-cor(score)
   rownames(temp)<-t(cats)
   colnames(temp)<-t(cats)
   pdf("test.pdf")
   plot(hclust(dist(temp)),hang=-1)
}
\author{
Alexander Griffith griffita@gmail.com
}

